refactor: 반복되는 코드 mapping으로 정리
